mkosi: add metavar=BOOL to boolean options
--help would print something like "--source-file-transfer-final SOURCE_FILE_TRANSFER_FINAL"
which takes a lot of space but is not very helpful. In particular it
might not be clear whether this expects some custom string or just a
yes/no boolean. Let's use "BOOL" instead to tell the user the type of
the argument, which immediately implies what values can be specified.

Similarly, say "--source-file-transfer METHOD", "--source-file-transfer-final METHOD".

Formatting is made a bit more consistent in the places that were touched.